Include versioning in docs config (#171)
* Use latest tag as build version

* include latest version in docs build
